BugFix - SJAKK-184 - Tournaments not showing in non-Chromium browsers
This commit will fix bug, where tournaments where not showing in non-chromium browsers.

The bug's origin was from the different browser engines' implementation of JSON.stringify. 
Chromium returns a string without ", but other returns  with. The Date class does not play well with the last one. 
The fix is to just give the Date constructor the toString() of the json object property, instead of using JSON.stringify()