Add NotFound component.
Add a component that we can route to when trying to route to non-existing paths.
Add routing to the not found component from components that takes props, and would still render. Solution was to add call to get the tounament, and if returned 404-status in response, route to not found component.