Fix several UI issues
The Dashboard button a organizer user sees was closer to the next tournament, so you could think the dashboard button was for the next tournament, not the one above.
Add margins to Tournaments component, so that its elements don't take up the whole width of the screen.
Change the width of tournament-items to be wider as parent element now is more narrow, which would make the tournament-items cramped and too small.
Add more margin between separate tournamentlists.
Add margin around Create tournament button so that  it is not clamped close to next element.